Added resolution and allowImageUpscaling in global app config
