+++ apps/renault-demo-app/src/pages/PhotoCapturePage/PhotoCapturePage.tsx
@@ -29,6 +29,8 @@ export function PhotoCapturePage() {
         sights={currentSights}
         onComplete={handleComplete}
         lang={i18n.language}
+        resolution={config.resolution}
+        allowImageUpscaling={config.allowImageUpscaling}
         format={config.format}
         quality={config.quality}
         tasksBySight={config.tasksBySight}
