+++ documentation/docs/configuration.md
@@ -39,6 +39,8 @@ The following table lists the available configuration options in the `CaptureApp
 | showCloseButton                    | `boolean`                                    | Indicates if the close button should be displayed in the HUD on top of the Camera preview.                                                                                                           |                                                   | `false`                     |
 | startTasksOnComplete               | `<code>boolean &#124; TaskName[]</code>`     | Value indicating if tasks should be started at the end of the inspection. See the `inspection-capture-web` package doc for more info.                                                                |                                                   | `true`                      |
 | tasksBySight                       | `Record<string, TaskName[]>`                 | Record associating each sight with a list of tasks to execute for it. If not provided, the default tasks of the sight will be used.                                                                  |                                                   |                             |
+| resolution                         | `CameraResolution`                           | Indicates the resolution of the pictures taken by the Camera.                                                                                                                                        |                                                   | `CameraResolution.UHD_4K`   |
+| allowImageUpscaling                | `boolean`                                    | Allow images to be scaled up if the device does not support the specified resolution in the `resolution` prop.                                                                                       |                                                   | `false`                     |
 | format                             | `CompressionFormat`                          | The output format of the compression.                                                                                                                                                                |                                                   | `CompressionFormat.JPEG`    |
 | quality                            | `number`                                     | Value indicating image quality for the compression output.                                                                                                                                           |                                                   | `0.8`                       |
 | allowSkipRetake                    | `boolean`                                    | If compliance is enabled, this prop indicate if the user is allowed to skip the retaking process if some pictures are not compliant.                                                                 |                                                   | `false`                     |
