+++ packages/camera-web/src/Camera/hooks/index.ts
@@ -4,4 +4,4 @@ export * from './useUserMedia';
 export * from './useCameraCanvas';
 export * from './useCompression';
 export * from './useTakePicture';
-export { type CanvasHandle, CameraFacingMode, CameraResolution, type CameraConfig } from './utils';
+export { type CanvasHandle, CameraFacingMode, type CameraConfig } from './utils';
