+++ packages/camera-web/src/Camera/index.ts
@@ -2,7 +2,6 @@ export { Camera, type CameraProps } from './Camera';
 export {
   type CameraConfig,
   CameraFacingMode,
-  CameraResolution,
   type UserMediaError,
   UserMediaErrorType,
 } from './hooks';
